feat: create person adapter service
