+++ src/app/services/person-adapter.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { PersonAdapterService } from './person-adapter.service';
+
+describe('PersonAdapterService', () => {
+  let service: PersonAdapterService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(PersonAdapterService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
